# openapi-mvp-v2.yaml
openapi: 3.0.3
info:
  title: Nomad API (MVP v2)
  version: 2.0.0
  description: |
    Expanded MVP API contract for Nomad with full hook coverage:
    - Auth + Onboarding restored
    - CRUD-expanded resources for courses, communities, posts, and messaging
servers:
  - url: http://localhost:3000/api
    description: Local dev
  - url: https://api.nomad.example.com
    description: Production (placeholder)
...
security:
  - bearerAuth: []
paths:
...
  /communities:
    get:
      tags: [Communities]
      summary: List communities
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: joined
          schema: { type: boolean }
      responses:
        "200":
          description: Communities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedCommunities" }
    post:
      tags: [Communities]
      summary: Create community
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCommunityRequest" }
      responses:
        "201":
          description: Created community
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Community" }
  /communities/{communityId}:
    get:
      tags: [Communities]
      summary: Get community
      parameters:
        - $ref: "#/components/parameters/CommunityId"
      responses:
        "200":
          description: Community
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Community" }
    patch:
      tags: [Communities]
      summary: Update community
      parameters:
        - $ref: "#/components/parameters/CommunityId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCommunityRequest" }
      responses:
        "200":
          description: Updated community
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Community" }
    delete:
      tags: [Communities]
      summary: Delete community
      parameters:
        - $ref: "#/components/parameters/CommunityId"
      responses:
        "204":
          description: Community deleted
  /communities/{communityId}/join:
    post:
      tags: [Communities]
      summary: Join community
      parameters:
        - $ref: "#/components/parameters/CommunityId"
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JoinState" }
    delete:
      tags: [Communities]
      summary: Leave community
      parameters:
        - $ref: "#/components/parameters/CommunityId"
      responses:
        "200":
          description: Left
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JoinState" }
...
  /calendar/events:
    get:
      tags: [Calendar]
      summary: List events
      parameters:
        - in: query
          name: start
          schema: { type: string, format: date-time }
        - in: query
          name: end
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedCalendarEvents" }
    post:
      tags: [Calendar]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCalendarEventRequest" }
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CalendarEvent" }
  /calendar/events/{eventId}:
    patch:
      tags: [Calendar]
      summary: Update event
      parameters:
        - in: path
          name: eventId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCalendarEventRequest" }
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CalendarEvent" }
    delete:
      tags: [Calendar]
      summary: Delete event
      parameters:
        - in: path
          name: eventId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Event deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CommunityId:
      in: path
      name: communityId
      required: true
      schema: { type: string }
...
  schemas:
...
    Community:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        banner: { type: string, nullable: true }
        color: { type: string }
        members: { type: integer }
        posts: { type: integer }
        tags:
          type: array
          items: { type: string }
        isJoined: { type: boolean }
        university: { type: string, nullable: true }
    CreateCommunityRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        banner: { type: string }
        color: { type: string }
        tags:
          type: array
          items: { type: string }
        university: { type: string }
    UpdateCommunityRequest:
      allOf:
        - $ref: "#/components/schemas/CreateCommunityRequest"
    JoinState:
      type: object
      properties:
        isJoined: { type: boolean }
...
    CalendarEvent:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        date: { type: string, format: date }
        startTime: { type: string }
        endTime: { type: string }
        location: { type: string, nullable: true }
        type:
          type: string
          enum: [class, study, social, deadline, custom]
        color: { type: string }
    CreateCalendarEventRequest:
      type: object
      properties:
        title: { type: string }
        date: { type: string, format: date }
        startTime: { type: string }
        endTime: { type: string }
        location: { type: string }
        type:
          type: string
          enum: [class, study, social, deadline, custom]
        color: { type: string }
    UpdateCalendarEventRequest:
      allOf:
        - $ref: "#/components/schemas/CreateCalendarEventRequest"
...
    PaginatedCommunities:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Community" }
        total: { type: integer }
...
    PaginatedCalendarEvents:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CalendarEvent" }
        total: { type: integer }
...